requirements.txt
# FastAPI backend
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.10.6

# Anthropic Claude API (replaces torch, transformers, spacy, nltk)
anthropic>=0.40.0

# PDF generation
fpdf2

# Optional utilities
python-multipart==0.0.6
